producer/requirements.txt
kafka-python
Flask==1.1.1